refactor: rewrite todolist, auth, task API, create files for enums,types, remove utills, create app & network errors. Create index.ts files for set short imports, fixed bugs, testing app
