+++ src/api/tasks-api.ts
@@ -1,0 +1,26 @@
+import { AxiosResponse } from 'axios'
+import { instance } from '../common'
+import { GetTasksResponse, ResponseType, TaskType, UpdateTaskModelType } from '../types'
+
+export const tasksAPI = {
+  getTasks(todolistId: string) {
+    return instance.get<GetTasksResponse>(`todo-lists/${todolistId}/tasks`)
+  },
+  deleteTask(todolistId: string, taskId: string) {
+    return instance.delete<ResponseType>(`todo-lists/${todolistId}/tasks/${taskId}`)
+  },
+  createTask(todolistId: string, title: string) {
+    return instance.post<
+      ResponseType<{ item: TaskType }>,
+      AxiosResponse<ResponseType<{ item: TaskType }>>,
+      { title: string }
+    >(`todo-lists/${todolistId}/tasks`, { title })
+  },
+  updateTask(todolistId: string, taskId: string, model: UpdateTaskModelType) {
+    return instance.put<
+      ResponseType<{ item: TaskType }>,
+      AxiosResponse<ResponseType<{ item: TaskType }>>,
+      UpdateTaskModelType
+    >(`todo-lists/${todolistId}/tasks/${taskId}`, model)
+  },
+}
