+++ src/api/todolists-api.ts
@@ -1,22 +1,7 @@
 import axios, { AxiosResponse } from 'axios'
 import { LoginDataType } from '../features/Login/Login'
-
-const instance = axios.create({
-  baseURL: 'https://social-network.samuraijs.com/api/1.1/',
-  withCredentials: true,
-})
-
-export class AuthAPI {
-  public login(data: LoginDataType) {
-    return instance.post<ResponseType<{ userId: number }>>('auth/login', data)
-  }
-  public logout() {
-    return instance.delete<ResponseType>('auth/logout')
-  }
-  public me() {
-    return instance.get<ResponseType<LoginType>>('auth/me')
-  }
-}
+import { instance } from '../common'
+import { ResponseType, TodolistType } from '../types'
 
 export const todolistsAPI = {
   getTodolists() {
@@ -38,86 +23,4 @@ export const todolistsAPI = {
       { title }
     )
   },
-  getTasks(todolistId: string) {
-    return instance.get<GetTasksResponse>(`todo-lists/${todolistId}/tasks`)
-  },
-  deleteTask(todolistId: string, taskId: string) {
-    return instance.delete<ResponseType>(`todo-lists/${todolistId}/tasks/${taskId}`)
-  },
-  createTask(todolistId: string, title: string) {
-    return instance.post<
-      ResponseType<{ item: TaskType }>,
-      AxiosResponse<ResponseType<{ item: TaskType }>>,
-      { title: string }
-    >(`todo-lists/${todolistId}/tasks`, { title })
-  },
-  updateTask(todolistId: string, taskId: string, model: UpdateTaskModelType) {
-    return instance.put<
-      ResponseType<{ item: TaskType }>,
-      AxiosResponse<ResponseType<{ item: TaskType }>>,
-      UpdateTaskModelType
-    >(`todo-lists/${todolistId}/tasks/${taskId}`, model)
-  },
 }
-
-export type LoginType = {
-  id: number
-  email: string
-  login: string
-}
-
-export type TodolistType = {
-  id: string
-  title: string
-  addedDate: string
-  order: number
-}
-export type ResponseType<D = {}> = {
-  resultCode: number
-  messages: Array<string>
-  fieldsErrors: Array<string>
-  data: D
-}
-
-export enum TaskStatuses {
-  New = 0,
-  InProgress = 1,
-  Completed = 2,
-  Draft = 3,
-}
-
-export enum TaskPriorities {
-  Low = 0,
-  Middle = 1,
-  Hi = 2,
-  Urgently = 3,
-  Later = 4,
-}
-
-export type TaskType = {
-  description: string
-  title: string
-  status: TaskStatuses
-  priority: TaskPriorities
-  startDate: string
-  deadline: string
-  id: string
-  todoListId: string
-  order: number
-  addedDate: string
-}
-export type UpdateTaskModelType = {
-  title: string
-  description: string
-  status: TaskStatuses
-  priority: TaskPriorities
-  startDate: string
-  deadline: string
-}
-type GetTasksResponse = {
-  error: string | null
-  totalCount: number
-  items: TaskType[]
-}
-
-export const authAPI = new AuthAPI()
