+++ src/common/common-instance.ts
@@ -1,0 +1,6 @@
+import axios from 'axios'
+
+export const instance = axios.create({
+  baseURL: 'https://social-network.samuraijs.com/api/1.1/',
+  withCredentials: true,
+})
