+++ src/features/TodolistsList/Todolist/Task/Task.tsx
@@ -1,40 +1,57 @@
 import React, { ChangeEvent, useCallback } from 'react'
 import { EditableSpan } from '../../../../components/EditableSpan/EditableSpan'
-import { TaskStatuses, TaskType } from '../../../../api/todolists-api'
-
-import { Delete } from '@mui/icons-material';
-import IconButton from '@mui/material/IconButton';
-import Checkbox from '@mui/material/Checkbox';
+import { Delete } from '@mui/icons-material'
+import IconButton from '@mui/material/IconButton'
+import Checkbox from '@mui/material/Checkbox'
+import { TaskStatuses } from '../../../../enums'
+import { TaskType } from '../../../../types'
 
 type TaskPropsType = {
-    task: TaskType
-    todolistId: string
-    changeTaskStatus: (id: string, status: TaskStatuses, todolistId: string) => void
-    changeTaskTitle: (taskId: string, newTitle: string, todolistId: string) => void
-    removeTask: (taskId: string, todolistId: string) => void
+  task: TaskType
+  todolistId: string
+  changeTaskStatus: (id: string, status: TaskStatuses, todolistId: string) => void
+  changeTaskTitle: (taskId: string, newTitle: string, todolistId: string) => void
+  removeTask: (taskId: string, todolistId: string) => void
 }
 export const Task = React.memo((props: TaskPropsType) => {
-    const onClickHandler = useCallback(() => props.removeTask(props.task.id, props.todolistId), [props.task.id, props.todolistId]);
+  const onClickHandler = useCallback(
+    () => props.removeTask(props.task.id, props.todolistId),
+    [props.task.id, props.todolistId]
+  )
 
-    const onChangeHandler = useCallback((e: ChangeEvent<HTMLInputElement>) => {
-        let newIsDoneValue = e.currentTarget.checked
-        props.changeTaskStatus(props.task.id, newIsDoneValue ? TaskStatuses.Completed : TaskStatuses.New, props.todolistId)
-    }, [props.task.id, props.todolistId]);
+  const onChangeHandler = useCallback(
+    (e: ChangeEvent<HTMLInputElement>) => {
+      let newIsDoneValue = e.currentTarget.checked
+      props.changeTaskStatus(
+        props.task.id,
+        newIsDoneValue ? TaskStatuses.Completed : TaskStatuses.New,
+        props.todolistId
+      )
+    },
+    [props.task.id, props.todolistId]
+  )
 
-    const onTitleChangeHandler = useCallback((newValue: string) => {
-        props.changeTaskTitle(props.task.id, newValue, props.todolistId)
-    }, [props.task.id, props.todolistId]);
+  const onTitleChangeHandler = useCallback(
+    (newValue: string) => {
+      props.changeTaskTitle(props.task.id, newValue, props.todolistId)
+    },
+    [props.task.id, props.todolistId]
+  )
 
-    return <div key={props.task.id} className={props.task.status === TaskStatuses.Completed ? 'is-done' : ''}>
-        <Checkbox
-            checked={props.task.status === TaskStatuses.Completed}
-            color="primary"
-            onChange={onChangeHandler}
-        />
+  return (
+    <div
+      key={props.task.id}
+      className={props.task.status === TaskStatuses.Completed ? 'is-done' : ''}>
+      <Checkbox
+        checked={props.task.status === TaskStatuses.Completed}
+        color="primary"
+        onChange={onChangeHandler}
+      />
 
-        <EditableSpan value={props.task.title} onChange={onTitleChangeHandler}/>
-        <IconButton onClick={onClickHandler}>
-            <Delete/>
-        </IconButton>
+      <EditableSpan value={props.task.title} onChange={onTitleChangeHandler} />
+      <IconButton onClick={onClickHandler}>
+        <Delete />
+      </IconButton>
     </div>
+  )
 })
