+++ src/features/TodolistsList/Todolist/Todolist.tsx
@@ -2,7 +2,6 @@ import React, { useCallback, useEffect } from 'react'
 import { AddItemForm } from '../../../components/AddItemForm/AddItemForm'
 import { EditableSpan } from '../../../components/EditableSpan/EditableSpan'
 import { Task } from './Task/Task'
-import { TaskStatuses, TaskType } from '../../../api/todolists-api'
 import { FilterValuesType } from '../todolists-reducer'
 import { fetchTasksTC } from '../tasks-reducer'
 import IconButton from '@mui/material/IconButton'
@@ -10,6 +9,8 @@ import Button from '@mui/material/Button'
 import { Delete } from '@mui/icons-material'
 import { useAppDispatch } from '../../../app/store'
 import { RequestStatusType } from '../../../app/app-reducer'
+import { TaskStatuses } from '../../../enums'
+import { TaskType } from '../../../types'
 
 type PropsType = {
   id: string
