+++ src/types/types.ts
@@ -1,0 +1,55 @@
+import { TaskPriorities, TaskStatuses } from '../enums'
+
+export type LoginType = {
+  id: number
+  email: string
+  login: string
+}
+
+export type TodolistType = {
+  id: string
+  title: string
+  addedDate: string
+  order: number
+}
+export type ResponseType<D = {}> = {
+  resultCode: number
+  messages: Array<string>
+  fieldsErrors: Array<string>
+  data: D
+}
+
+export type TaskType = {
+  description: string
+  title: string
+  status: TaskStatuses
+  priority: TaskPriorities
+  startDate: string
+  deadline: string
+  id: string
+  todoListId: string
+  order: number
+  addedDate: string
+}
+export type UpdateTaskModelType = {
+  title: string
+  description: string
+  status: TaskStatuses
+  priority: TaskPriorities
+  startDate: string
+  deadline: string
+}
+export type GetTasksResponse = {
+  error: string | null
+  totalCount: number
+  items: TaskType[]
+}
+
+export type UpdateDomainTaskModelType = {
+  title?: string
+  description?: string
+  status?: TaskStatuses
+  priority?: TaskPriorities
+  startDate?: string
+  deadline?: string
+}
