+++ src/utills/handle-server-network-error.ts
@@ -1,0 +1,7 @@
+import { Dispatch } from 'redux'
+import { AppActionsType, setErrorAC, setStatusAC } from '../app/app-reducer'
+
+export const handleServerNetworkError = (dispatch: Dispatch<AppActionsType>, error: string) => {
+  dispatch(setErrorAC(error))
+  dispatch(setStatusAC('failed'))
+}
